Switch to Rocket from Express
Under advisement from Kira I am switching to using Rocket instead of Express
so that searches do not enter the command line at all. Doing this I also
managed to use opencv.js so that the only data being transmitted between the
server and client is JSON data. This helps improve security because I do not
need to ever put files on disk or use the command line.

There are some rough spots, but that is due to my lack of knowledge of rust
and threads. I hope to improve these and make them look cleaner in the future